Added new animations, removed old Anim BP's, Added Detection multiplier
Added all of the new animations, all with their own animBP, i also added
the detection multiplier at its basic form, need to hook up heavy and
medium animations but the multiplier and kill timer are in. - Also
changed level 1 so it worked with the elevator and the post processing
volume.